Silence a bunch of MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@23033 f0d54148-0727-0410-94bb-9a71ac55c965

+++ alg/gdalcutline.cpp
@@ -174,7 +174,7 @@ BlendMaskGenerator( int nXOff, int nYOff, int nXSize, int nYSize,
                 dfRatio = 0.5 + (dfDist / dfBlendDist) * 0.5;
             }                
 
-            pafValidityMask[iX + iY * nXSize] *= dfRatio;
+            pafValidityMask[iX + iY * nXSize] *= (float)dfRatio;
         }
     }
